Thiết lập ràng buộc Categories - Courses - Teacher

+++ Modules/Courses/database/migrations/2025_11_26_130603_create_courses_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->text('detail')->nullable();
 
             // Quan trọng – Cho phép null
-            $table->unsignedBigInteger('teacher_id')->nullable();
+            $table->unsignedInteger('teacher_id')->nullable();
 
             $table->string('thumbnail')->nullable();
